Split up the code for the animedatabase commands
