+++ xenian/bot/commands/animedatabase_utils/post.py
@@ -1,0 +1,99 @@
+import os
+from typing import IO
+
+from telegram import InputMediaPhoto, PhotoSize
+
+
+class Post:
+    class PostDictWrapper:
+        post = {}
+
+        def __getattribute__(self, item):
+            return self.post.get('')
+
+    def __init__(self, post: dict, media: str or IO or PhotoSize = None, caption: str = None, post_url: str = None):
+        self.post = post
+        self.post_url = post_url
+        self._telegram = InputMediaPhoto(media, caption)
+        self._file = None
+
+    def is_image(self, include_gif: bool = False) -> bool or None:
+        if self.file_extension is None:
+            return
+
+        image_extensions = ['jpg', 'jpeg', 'png']
+        if include_gif:
+            image_extensions.append('gif')
+        return self.file_extension in image_extensions
+
+    def is_video(self, include_gif: bool = True) -> bool or None:
+        if self.file_extension is None:
+            return
+
+        video_extensions = ['webm', 'mp4']
+        if include_gif:
+            video_extensions.append('gif')
+        return self.file_extension in video_extensions
+
+    @property
+    def file_extension(self):
+        media = self.media
+        if isinstance(media, PhotoSize):
+            if self._file is None:
+                self._file = media.get_file()
+            media = self._file.file_path
+
+        if isinstance(media, str):
+            split = os.path.splitext(self.media)
+            if len(split) != 2:
+                return
+            return split[1]
+        return
+
+    @property
+    def media(self) -> str or IO or PhotoSize:
+        return self.telegram.media
+
+    @media.setter
+    def media(self, value: str or IO or PhotoSize):
+        self.telegram.media = value
+
+    @property
+    def caption(self) -> str:
+        return self.telegram.caption
+
+    @caption.setter
+    def caption(self, value: str):
+        self.telegram.caption = value
+
+    @property
+    def telegram(self) -> InputMediaPhoto:
+        return self._telegram
+
+    @telegram.setter
+    def telegram(self, value: InputMediaPhoto or str or tuple or list):
+        actual_value = None
+        if isinstance(value, InputMediaPhoto):
+            actual_value = value
+        elif (isinstance(value, tuple) or isinstance(value, list)) and len(value) == 2:
+            actual_value = InputMediaPhoto(*value)
+        elif isinstance(value, str):
+            if self._telegram:
+                actual_value = self._telegram
+                actual_value.media = value
+            else:
+                actual_value = InputMediaPhoto(value)
+        if not actual_value:
+            raise ValueError('Value must either be a tuple/list, dict, InputMediaPhoto or str')
+
+        self._telegram = actual_value
+
+
+class PostError(Exception):
+    IMAGE_NOT_FOUND = 0
+    WRONG_FILE_TYPE = 10
+    UNDEFINED_ERROR = 20
+
+    def __init__(self, code: int, post: Post = None):
+        self.code = code
+        self.post = post
