+++ xenian/bot/commands/animedatabases.py
@@ -4,319 +4,25 @@ import re
 import zipfile
 from collections import OrderedDict
 from copy import deepcopy
-from typing import Any, Callable, IO, Iterable
+from typing import Any, Callable, Iterable
 
 import requests
-from pybooru import Danbooru as PyDanbooru, Moebooru as PyMoebooru
-from pybooru.resources import SITE_LIST
 from requests.exceptions import MissingSchema
-from requests_html import HTMLSession
-from telegram import Bot, ChatAction, InputFile, InputMediaPhoto, Message, PhotoSize, Update
-from telegram.error import BadRequest, NetworkError, TimedOut
+from telegram import Bot, ChatAction, InputFile, InputMediaPhoto, Update
 from telegram.ext import run_async
 
 from xenian.bot import mongodb_database
+from xenian.bot.commands.animedatabase_utils.base_service import BaseService
+from xenian.bot.commands.animedatabase_utils.danbooru_service import DanbooruService
+from xenian.bot.commands.animedatabase_utils.message_queue import MessageQueue
+from xenian.bot.commands.animedatabase_utils.moebooru_service import MoebooruService
+from xenian.bot.commands.animedatabase_utils.post import Post, PostError
 from xenian.bot.settings import ANIME_SERVICES
 from xenian.bot.utils import CustomNamedTemporaryFile, TelegramProgressBar, download_file_from_url_and_upload
 from xenian.bot.utils.telegram import retry_command
 from . import BaseCommand
 
 __all__ = ['animedatabases']
-
-SITE_LIST['safebooru'] = {'url': 'https://safebooru.donmai.us'}
-
-
-class Post:
-    class PostDictWrapper:
-        post = {}
-
-        def __getattribute__(self, item):
-            return self.post.get('')
-
-    def __init__(self, post: dict, media: str or IO or PhotoSize = None, caption: str = None, post_url: str = None):
-        self.post = post
-        self.post_url = post_url
-        self._telegram = InputMediaPhoto(media, caption)
-        self._file = None
-
-    def is_image(self, include_gif: bool = False) -> bool or None:
-        if self.file_extension is None:
-            return
-
-        image_extensions = ['jpg', 'jpeg', 'png']
-        if include_gif:
-            image_extensions.append('gif')
-        return self.file_extension in image_extensions
-
-    def is_video(self, include_gif: bool = True) -> bool or None:
-        if self.file_extension is None:
-            return
-
-        video_extensions = ['webm', 'mp4']
-        if include_gif:
-            video_extensions.append('gif')
-        return self.file_extension in video_extensions
-
-    @property
-    def file_extension(self):
-        media = self.media
-        if isinstance(media, PhotoSize):
-            if self._file is None:
-                self._file = media.get_file()
-            media = self._file.file_path
-
-        if isinstance(media, str):
-            split = os.path.splitext(self.media)
-            if len(split) != 2:
-                return
-            return split[1]
-        return
-
-    @property
-    def media(self) -> str or IO or PhotoSize:
-        return self.telegram.media
-
-    @media.setter
-    def media(self, value: str or IO or PhotoSize):
-        self.telegram.media = value
-
-    @property
-    def caption(self) -> str:
-        return self.telegram.caption
-
-    @caption.setter
-    def caption(self, value: str):
-        self.telegram.caption = value
-
-    @property
-    def telegram(self) -> InputMediaPhoto:
-        return self._telegram
-
-    @telegram.setter
-    def telegram(self, value: InputMediaPhoto or str or tuple or list):
-        actual_value = None
-        if isinstance(value, InputMediaPhoto):
-            actual_value = value
-        elif (isinstance(value, tuple) or isinstance(value, list)) and len(value) == 2:
-            actual_value = InputMediaPhoto(*value)
-        elif isinstance(value, str):
-            if self._telegram:
-                actual_value = self._telegram
-                actual_value.media = value
-            else:
-                actual_value = InputMediaPhoto(value)
-        if not actual_value:
-            raise ValueError('Value must either be a tuple/list, dict, InputMediaPhoto or str')
-
-        self._telegram = actual_value
-
-
-class PostError(Exception):
-    IMAGE_NOT_FOUND = 0
-    WRONG_FILE_TYPE = 10
-    UNDEFINED_ERROR = 20
-
-    def __init__(self, code: int, post: Post = None):
-        self.code = code
-        self.post = post
-
-
-class BaseService:
-    type = 'base'
-
-    def __init__(self, name: str, url: str, api: str = None, username: str = None, password: str = None):
-        self.name = name
-        self.url = url.lstrip('/') if url is not None else None
-        self.api = api
-        self.username = username
-        self.password = password
-
-        self.count_qualifiers_as_tag = False
-        self.client = None
-        self.session = None
-        self.tag_limit = None
-        self.censored_tags = None
-
-    def init_client(self):
-        raise NotImplemented
-
-    def init_session(self):
-        raise NotImplemented
-
-
-class DanbooruService(BaseService):
-    FREE_LEVEL = 20
-    GOLD_LEVEL = 30
-    PLATINUM_LEVEL = 31
-    BUILDER = 32
-    JANITOR = 35
-    MODERATOR = 40
-    ADMIN = 50
-
-    LEVEL_RESTRICTIONS = {
-        'tag_limit': {
-            FREE_LEVEL: 2,
-            GOLD_LEVEL: 6,
-            PLATINUM_LEVEL: 12,
-            BUILDER: 32,
-            JANITOR: 35,
-            MODERATOR: 40,
-            ADMIN: 50,
-        },
-        'censored_tags': {
-            FREE_LEVEL: True,
-            GOLD_LEVEL: False,
-            PLATINUM_LEVEL: False,
-            BUILDER: False,
-            JANITOR: False,
-            MODERATOR: False,
-            ADMIN: False,
-        }
-    }
-
-    type = 'danbooru'
-
-    def __init__(self, name: str, url: str, api: str = None, username: str = None, password: str = None) -> None:
-        super(DanbooruService, self).__init__(name=name, url=url, api=api, username=username, password=password)
-        self.user_level = None
-
-        self.init_client()
-        self.init_session()
-
-    def init_client(self):
-        if self.api:
-            if not self.username:
-                raise ValueError('Danbooru API Services need a Username when API key is given.')
-            self.client = PyDanbooru(site_name=self.name, site_url=self.url, api_key=self.api, username=self.username)
-        else:
-            self.client = PyDanbooru(site_name=self.name, site_url=self.url)
-
-        self.user_level = self.get_user_level()
-        self.tag_limit = self.LEVEL_RESTRICTIONS['tag_limit'][self.user_level]
-        self.censored_tags = self.LEVEL_RESTRICTIONS['censored_tags'][self.user_level]
-
-        if not self.url:
-            self.url = self.client.site_url.lstrip('/')
-
-    def init_session(self):
-        if self.username and self.password and self.url:
-            self.session = HTMLSession()
-            login_page = self.session.get(f'{self.url.lstrip("/")}/session/new')
-            form = login_page.html.find('.simple_form')[0]
-
-            login_data = {
-                'name': self.username,
-                'password': self.password,
-                'remember': '1',
-            }
-            for input in form.find('input'):
-                value = input.attrs.get('value', None)
-                name = input.attrs.get('name', None)
-                if name:
-                    login_data.setdefault(name, value)
-
-            self.session.post(f'{self.url.lstrip("/")}/session', login_data)
-
-    def get_user_level(self):
-        user_level = 20
-        if self.username:
-            user = self.client.user_list(name_matches=self.client.username)
-            user_level = user[0]['level']
-        return user_level
-
-
-class MoebooruService(BaseService):
-    type = 'moebooru'
-
-    def __init__(self, name: str, url: str, username: str = None, password: str = None,
-                 hashed_string: str = None) -> None:
-        super(MoebooruService, self).__init__(name=name, url=url, username=username, password=password)
-        self.tag_limit = 6
-        self.hashed_string = hashed_string
-        self.count_qualifiers_as_tag = True
-
-        self.init_client()
-
-    def init_client(self):
-        if self.username and self.password:
-            self.client = PyMoebooru(site_name=self.name, site_url=self.url, hash_string=self.hashed_string,
-                                     username=self.username, password=self.password)
-            return
-
-        self.client = PyMoebooru(site_name=self.name, site_url=self.url)
-        if not self.url:
-            self.url = self.client.site_url.lstrip('/')
-
-
-class MessageQueue:
-    tag_limit = 6
-
-    def __init__(self, total: int, message: Message, group_size: int = None):
-        self.total = total
-        self.message = message
-        self.group_size = group_size or 1
-
-        self.sent = 0
-        self.errors = []
-
-    def report(self, error: PostError = None):
-        if error:
-            self.errors.append(error)
-
-        self.sent += 1
-        if self.sent == self.total:
-            self.finished()
-
-    def finished(self):
-        lines = []
-        if self.message.chat.type not in ['group', 'supergroup']:
-            lines.append('Images has been sent')
-
-        if self.errors:
-            error_codes = set(error.code for error in self.errors)
-            if PostError.IMAGE_NOT_FOUND in error_codes:
-                lines.append('Some files could not be retrieved')
-                for error in filter(lambda e: e.code == PostError.IMAGE_NOT_FOUND and e.post.post_url, self.errors):
-                    lines.append(f'- {error.post.post_url}')
-
-            if PostError.WRONG_FILE_TYPE in error_codes and self.group_size > 1:
-                lines.append('Videos were skipped because they cannot be sent via a group.')
-                for error in filter(lambda e: e.code == PostError.WRONG_FILE_TYPE and e.post.post_url, self.errors):
-                    lines.append(f'- {error.post.post_url}')
-
-            if PostError.UNDEFINED_ERROR in error_codes:
-                lines.append('Some files could not be sent')
-
-        if lines:
-            self.message.reply_text('\n'.join(lines),
-                                    reply_to_message_id=self.message.message_id,
-                                    disable_web_page_preview=True)
-
-    @staticmethod
-    def message_queue_exc_handler(argument_name: str):
-        def wrapper(func, *args, **kwargs):
-            def inner_wrapper(*args, **kwargs):
-                queue = kwargs.get(argument_name)
-                if not queue:
-                    queues = [arg for arg in args if isinstance(arg, MessageQueue)]
-                    if len(queues) > 1:
-                        raise AttributeError('Got more than one MessageQueue\'s')
-                    elif queues:
-                        queue = queues[0]
-                if not queue:
-                    raise AttributeError('No MessageQueue found')
-                try:
-                    return func(*args, **kwargs)
-                except (BadRequest, TimedOut, NetworkError) as e:
-                    if isinstance(e, NetworkError) and 'The write operation timed out' not in e.message:
-                        raise e
-
-                    for item in range(queue.group_size):
-                        queue.report(PostError(code=PostError.UNDEFINED_ERROR))
-
-            return inner_wrapper
-
-        return wrapper
 
 
 class AnimeDatabases(BaseCommand):
